Bugfixes W. Huber, R. Gentleman 23 Feb 2003
affy/src/read_cdffile.c:
=========================
In readline_affy_file(): returns NULL if EOF. It did not before,
which led to an infinite loop in some cases.

In getProperty():
Fixed the while loop over the reading in of the file such that will
actually terminate if the search string does not occur in the file

--> The same probably needs to be fixed in other places as well <--
e.g. goToUnit(), goToUnitStartingWith()

affy/R/read.affybatch.R
=========================
replaced the line
  if( dim(intensity(cel)) != dim.intensity )
by
  if( !all(dim(intensity(cel)) == dim.intensity) )
since the objects that are compared are vectors and 'if' expects a scalar