Oops.  Adding data files that were suppposed to be part of the previous commit.
[old log message follows]

Rework how AnnBuilder gets the source data URLs.  Changed to a plain
text format.  Actual public domain source data URLs moved to
pubDataURLs.  AnnBuilder will use the URLs listed in sourceURLs during
package building.  This is to make it easy to create a local mirror of
public domain data and avoid repeated downloads of the same data.

The downloadSourceData.R contains a function to create a local mirror
of the public data.  Except for KEGG which must be downloaded as a
tar.gz and untarred manually.