Version bump to begin Bioc 1.8 DEVEL work.
This version bump is unusual in that it is part of a catch-up version bump.
The project has decided to work with an even/odd scheme.

All versions are x.y.z.  
In release y is even.
In devel y is odd.

This change sets everyone to the next odd y (next as compared to
version in 1.7 release).