Fixed bug in Makesense so it respects subsetted exprSets
The code was expecting exprSets containing _every_ probe on the array
and failed for exprSets that have been subsetted, for example, gone
through on-specific filtering.  This patch corrects that along with a
few other minor fixes, for example, we no longer mandate human
chromosomes and trust the datapkgCHR environment to contain reasonable
things.

Default color for plotChr changed blue => black.