Added dependencies: BufferedMatrix, BufferedMatrixMethods
Added capabilities to use BufferedMatrix objects in assayData
of *FeatureSet objects (only for Affymetrix right now).

Added read.celfiles2() and rma2(): these functions work directly
with BufferedMatrix objects, saving memory. My goal is to
replace the current functions by these. Already tested and validated.

Modified pm(), so it returns a BufferedMatrix. This allows the use of
rma2().

Changed rma(), so it coerces the result of pm() to matrix.