Import AnnotationDbi, fix lookUp to work with SQLite-based pkgs
This patch removes the call to require and the check for the env
in the data package environment.  These aren't that useful since:
(1) code that depends upon a data package should require it itself,
and (2) we end up just doing get() on the object name so we aren't
protected from masking issues in the first place.

With these changes, you can load SQLite-based packages, such as
hgu95v2db and use lookUp with it (you still specify "hgu95av2", not
"hgu95av2db").