Another graphNEL => graphAM optimization
Here we avoid much copying of large data by sneaking around the
graphAM initialize method.  This is less than ideal from a design
perspective since we don't get the same level of checking and
encapsulation.  But this gives a big improvement and the correct
result so it seems worth it for now.