Update S4 objects in rda files
I used a script to find serialized S4 instances that appeared to have
been created with an R before the S4SXP was in place.

Biobase::updateObject was then used to update the object and a new rda
file was written.  All packages passed check after the update, but
maintainers should review the changes.