Rename class Untyped => NullIdentifier
I don't think this class will be useful other than as a null object
placeholder.  I don't like Untyped as it is too generic.  Since the
other subclasses have "Identifier" as part of the name, I think this
one should as well.  I'm not too worried about conflict with the NULL
database.