Update DB_PATH in .onLoad for pd.mapping template
If the user has different versions installed in multiple package
libraries, we need to get the right one at runtime.  Note that for the
lazyloading mechanism to work and so that we can create a valid map
object, we also need the path in package scope for the initial
install.