Changes and bug fixes for making chip based pkgs.
These changes and bug fixes allow things to work properly for making
chip based databases and bundling these into packages.  There is still
a minor issue with arabidopsis to fix. And then the intermediate
databases need to be bundled into packages for easy distribution and
the default paths set for these helper functions.