Adds a vignette for SQLForge, and some bug fixes.
This checkin adds a vignette for the newer SQLForge functions as well
as some bug fixes for some of the populate functions so that they can
correctly build their databases in the specified directories.  Also, a
much shorter example file was added to cut down on the time needed to
check the new vignette.