Fixed bad but subtle bug in SQLforge.
This fixes a bad bug in SQLforge that would under certain rare
circumstances, map the wrong type of indentifier into the entrez gene
field.  Now the code does a better job of making sure that when it has
an alternate ID, to find a correct entrez gene id if one can be found
and stick that into the entrez gene field.