Fix to try and clear up ".db" name inconsistencies.
This adds some code to take the name that users put for the
"annotation" of a param and then parse that so that it always removes
the ".db" extension (if they had bothered to type it).  This
simplifies the internal representation and allows us to not have to
remove ".db" from various places in the code.  It also should help
clear up confusion for users about whether they need to type
hgu95av2.db or hgu95av2 when using the parameter objects since either
will now work the same.  Finally, the shorter name is the true prefix
for the package since that is what is used internally to name all the
mappings.  Leaving the ".db" around just creates more work for
everyone to clean up via string substitution.  Of course if the world
suddenly breaks tomorrow then we will have to reconsider...