Just a quick fix to restore functionality to example(getGene)
Inside checkWrapperArgs(), martFilters(mart) is now returning
something other than what it used to.  Now its a dataframe which is
not easily coerced to a list.  listFilters() still works the same
though and can also work here in our pinch.  So that is what it does
for now to restore functionality.