Biostrings 2.15.12
The vmatchPDict() family now supports a dictionary of patterns that has not
been preprocessed (i.e. the 'pdict' arg can receive an XStringSet object).
(For completeness, I still need to do the same for the matchPDict() family.)

When using a non-preprocessed dictionary for fuzzy matching, there are no
Trusted Band constraints anymore, i.e. the user-supplied 'max.mismatch',
'min.mismatch' and 'fixed' values define a search criteria that is exactly
the same as if matchPattern() or vmatchPattern() were used in a loop.
However, using vmatchPDict() on a non-preprocessed dictionary can be slightly
faster than using matchPattern() or vmatchPattern() in a "lapply loop",
because, with vmatchPDict(), looping is done at the C-level.