Fixes warnings.
This package had warnings because the manual pages for partially
documented internal functions were incomplete.  These functions were
not exported, the only "crime" was that the author had bothered to
write part of a manual page about them.  It doesn't make sense to get
rid of the partial manual pages, or to complete the documentation
either (it's really not necessary for me to guess what the author
intended when the function is not even exported).  So I added a
comment to these arguments tha the function in question was an
internal function and therefore some arguments may not have been
documented.