Bug fix for SQLForge
This change just removes the behavior where SQLForge was altering the
installed .db0 packages in innocuous ways when building a
package. (VACUUM seems to affect the attached packages, and stuff was
being inserted into (and then removed) from the metadata tables)
SQLForge is not allowed to alter these files in ANY way, so this gets
rid of these scary looking behaviors.