begin commening the 'normal' matrix as this is not used.
Parts of the 'normal' matrix appear to be used in fit.lm4 -- the
wrapper for nonpolymorphic markers on X. Testing here will be needed
once these lines are commented.

Should snpflags and normal be part of the CNSet object an instantiated
upon initialization?  I think so.  snpflags need only be the same
dimension as the parameters and can be included in the
LinerModelParameter class.