Changes the checking mechanism for findOverlaps
Instead of guessing what the user intends, findOverlaps will now just
tell the user a warning if their sequence vectors are not the same.
If the vectors are identical, then things are obviously ok.  And if
one of the two vectors is contained in the other, things are also ok.
But if both vectors are only partially overlapping, users will get a
warning.