Adds circ_seqs argument to makeTranscriptDbFromUCSC and makeTranscriptDbFromBiomart
This adds the circ_seq argument to these two methods along with two
discovery methods to see what chromosomes are being pulled from a
given source.  Users are expected to pass in a character string to
indicate which things they want to be marked as circular.  The
discovery methods allow the user to see what strings they can choose
from.  A default character vector is also provided, exposed, and
documented to allow users to automatically mark most mitochondrial
sequences as circular, If there are other sequences that need to be so
marked, users can append to this default vector.