Finally exposes rudimentary functionality for generating org pkgs
The goal is to be able to generate an org package on the fly.  
This code should start paving the way.  
I still need to add code for creating the GO tables indices, 
and metadata, along with code to wrap the DB into a package.  
Also, there are some bugs that still need to be adressed.   
Too many EGs, can cause you to run out of memory, and there 
seems to be an issue with the XML from NCBI for retrieving 
the Entrez Genes from a tax ID.