Added S3 predict method for classifierOutput instances.
Adding predict.classifierOutput required adding a predicter slot to the 
learnerSchema class. The default predicter function is defined in the 
class prototype a simply calls the specific predict function for a newdata 
object. Specific functions can be defined in the MLIPredicters.R file. 
Currently, a list is returned, but this will probably be updated to a 
classiferOutput (or similar) instance. Wrote predict.classifierOutput.Rd 
and updated MLearn-new.Rd to reflect changes in makeLearnerSchema.

Also moved plsda-related code from plsdaI.R to other R files.