Comments out a bad example in makeProbePackage.  :(
This package MUST get pushed and I found a strange 
problem in the makeProbePackage vignette.  So I am 
taking an extreme action and setting eval to FALSE 
for tonight only.  

There might be something funky happening in 
makeProbePackage which I will notify Herve about 
immediately, and which will have to be fixed tomorrow.

In the meantime a version number check for this package 
is breaking everything else and it just does not 
need to be doing so.