Various housekeeping changes, mostly intended to keep the site forward-compatible with newer versions of nanoc and ruby.
Also, add a config variable that tracks which repositories of devel are
available. When we do a release, typically the new devel repos 
(bioc, data/annotation, and data/experiment) show up one at a time. 

Nothing is yet done with this variable but it will eventually avoid
a lot of the hacking that happens with new releases.