edgeR 3.0.4
adjustedProfLik() is more efficient (avoiding an unnecessary call to predFC), thus making dispersion estimation for glm models faster.

In glmFit(), the maximum number of iterations in the call to mglmLevenberg() is reduced to 250.  Previously the number of iterations was so large that the Levenberg damping parameter could become exactly zero due to floating point underflow, making an infinite loop possible in rare pathological cases.